requirements.txt
datasets # Add -U flag
pillow
matplotlib
numpy
openai
anthropic
notebook
